M8: Rename validation test files; move state-probs tests to test_regime.py
Three renames driven by Phase 1's src moves:
- `tests/test_error_handling_invalid_vf.py` → `tests/test_invalid_vf.py`
  (the "error_handling" concept is gone from src).
- `tests/test_validate_array_indexing.py` → `tests/test_ast_inspection.py`
  (AST helpers moved to `lcm/utils/ast_inspection.py`).
- Extract the four `validate_transition_probs` state-probs tests from
  `tests/test_pandas_utils.py` into `tests/test_regime.py` (the
  function now lives in `lcm/user_regime.py`). Duplicate the
  three-line `_make_partner_probs_array` helper into the new
  location rather than imposing a cross-file import.

Closes the Phase 1 plan in `Phase 1 — Validation Cleanup.md`.
